predict-sales-executive/requirements.txt
flask>=3.0.0
kafka-python>=2.0.2
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
